* ConjureItem fixes * Fix warnings

+++ otserv/branches/avesta/party.cpp
@@ -341,7 +341,7 @@ bool Party::setSharedExperience(Player* player, bool _sharedExpActive)
 
 void Party::shareExperience(uint32_t experience)
 {
-	uint32_t shareExperience = std::ceil(((float)experience / (memberList.size() + 1)));
+	uint32_t shareExperience = (uint32_t)std::ceil(((float)experience / (memberList.size() + 1)));
 	for(PlayerVector::iterator it = memberList.begin(); it != memberList.end(); ++it){
 		(*it)->onGainSharedExperience(shareExperience);
 	}
